+++ lib/thermite/rails/tasks/project_thermite_test_task.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require_relative 'project_rake_task'
+
+module Thermite
+  module Rails
+    module Tasks
+      class ProjectThermiteTestTask < ProjectRakeTask
+        # @return [String]
+        def task_name
+          "thermite:test:#{crate_name_for_ruby}"
+        end
+
+        def define_rake_task
+          desc "Using thermite, (cargo) test crate #{crate_name}"
+          task(task_name) { run_task('thermite:test') }
+        end
+      end
+    end
+  end
+end