+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.y.z] - 2018-mm-dd
+
+### Improvements
+
+* Moved `*Task` classes to `lib/thermite/rails/tasks/`.
+
 ## [0.1.0] - 2018-03-27
 
 ### Improvements